% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalKinetiX.R
\name{CausalKinetiX}
\alias{CausalKinetiX}
\title{CausalKinetix}
\usage{
CausalKinetiX(X)
}
\arguments{
\item{X}{}
}
\value{
object of class 'CausalKinetiX' consisting of the following
  elements

\item{theta}{parameters}
}
\description{
CausalKinetiX
}
\details{
For further details see the references.
}
\examples{

x <- 4
}
\references{
Pfister, N., S. Bauer, J. Peters (2018).
Identifying Causal Structure in Large-Scale Kinetic Systems
ArXiv e-prints (arXiv:1810.11776).
}
\seealso{

}
\author{
Niklas Pfister, Stefan Bauer and Jonas Peters
}
