% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.hidden.R
\name{generate.data.hidden}
\alias{generate.data.hidden}
\title{Hidden variable model}
\usage{
generate.data.hidden(env = rep(1, 10), L = 15, par.noise = list(noise
  = 0.01, only.target.noise = TRUE, relativ = FALSE),
  intervention = "initial_blockreactions5", hidden = TRUE,
  intervention.par = NA, ode.solver = "lsoda", seed = NA,
  silent = FALSE)
}
\arguments{
\item{env}{integer vector of length n encoding to which experiment
each repetition belongs.}

\item{L}{number of time points for evaluation.}

\item{par.noise}{list of parameters that specify the added noise.}

\item{intervention}{type of intervention.}

\item{hidden}{boolean whether the variables H1 and H2 should be
removed from output.}

\item{intervention.par}{parameters used in the interventions.}

\item{ode.solver}{specifies which ODE solver to use when solving
ODE.}

\item{seed}{random seed.}

\item{silent}{set to FALSE if status output should be produced.}
}
\value{
list consisting of the following elements

\item{simulated.data}{D-matrix of noisy data.}
\item{time}{vector containing time points}
\item{env}{vector specifying the experimental environment.}
\item{simulated.model}{object returned by ODE solver.}
\item{true.model}{vector specifying the target equation model.}
\item{target}{target variable.}
}
\description{
Generate sample data from the hidden variable model.
}
\details{
For further details see the references.
}
\examples{

simulation.obj <- generate.data.hidden(env=rep(1:5, 3),
                                       L=15,
                                       par.noise=list(noise=0.02,
                                                      only.target.noise=FALSE,
                                                      relativ=TRUE),
                                       intervention="initial_blockreactions5",
                                       intervention.par=0.2)

D <- simulation.obj$simulated.data
fulldata <- simulation.obj$simulated.model
time <- simulation.obj$time
plot(fulldata[[1]][,1], fulldata[[1]][,2], type= "l", lty=2,
     xlab="time", ylab="concentration")
points(time, D[1,1:length(time)], col="red", pch=19)
legend("topright", c("true trajectory", "observations"),
       col=c("black", "red"), lty=c(2, NA), pch=c(NA, 19))
}
\references{
Pfister, N., S. Bauer, J. Peters (2018).
Identifying Causal Structure in Large-Scale Kinetic Systems
ArXiv e-prints (arXiv:1810.11776).
}
\seealso{
The functions \code{\link{generate.data.maillard}} and
  \code{\link{generate.data.targetmodel}} allow to simulate ODE
  data from two additional models.
}
\author{
Niklas Pfister, Stefan Bauer and Jonas Peters
}
