% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalKinetiX.modelranking.R
\name{CausalKinetiX.modelranking}
\alias{CausalKinetiX.modelranking}
\title{CausalKinetix.modelranking}
\usage{
CausalKinetiX.modelranking(D, times, env, target, models, pars)
}
\arguments{
\item{D}{data matrix. Should have dimension n x (L*d), where n is
the number of repetitions (over all experiments), L is the
number of time points and d is the number of predictor
variables.}

\item{times}{vector of length L specifying the time points at
which data was observed.}

\item{env}{integer vector of length n encoding to which experiment
each repetition belongs.}

\item{target}{integer specifing which variable is the target.}

\item{models}{list of models. Each model is specified by a list of
vectors specifiying the variables included in the interactions
of each term.}

\item{pars}{list of the following parameters: \code{pen.degree}
(default 2) specifies the penalization degree in the smoothing
spline, \code{num.folds} (default 2) number of folds used in
cross-validation of smoothing spline, \code{include.vars}
(default NA) specifies variables that should be included in each
model, \code{include.intercept} (default FALSE) specifies
whether to include a intercept in models, \code{pooling}
(default FALSE) specifies whether to pool repetitions in each
environment, \code{smoothing} (default FALSE) specifies whether
to smooth data observations before fitting, \code{smooth.Y}
(default FALSE) specifies whether to smooth target observations
before fitting, \code{regression.class} (default OLS) other
options are signed.OLS, optim, random.forest,
\code{sample.splitting} (default "loo") either leave-one-out
(loo) or no splitting (none), \code{score.type} (default "mean")
specifies the type of score funtion to use,
\code{integrated.model} (default TRUE) specifies whether to fit
the integrated or the derived model, \code{splitting.env}
(default NA) an additonal environment vector used for scoring,
\code{weight.vec} (default rep(1, length(env)) a weight vector
used in the scoring, \code{set.initial} (default FALSE)
specifies whether to fix the initial value, \code{silent}
(default TRUE) turn of additional output, \code{show.plot}
(default FALSE) show diagnostic plots.}
}
\value{
returns a vector with the same length as models containing
  the stability scores
}
\description{
Applies CausalKinetiX framework to rank a list models according to
their stability.
}
\details{
This function only scores specified models and does not include a
variable ranking.
}
\examples{

x <- 4
}
\references{
Pfister, N., S. Bauer, J. Peters (2018).
Identifying Causal Structure in Large-Scale Kinetic Systems
ArXiv e-prints (arXiv:1810.11776).
}
\seealso{
The function \code{\link{CausalKinetiX}} is a wrapper for
  this function that also computes the variable ranking.
}
\author{
Niklas Pfister, Stefan Bauer and Jonas Peters
}
