% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalKinetiX.modelranking.R
\name{CausalKinetiX.modelranking}
\alias{CausalKinetiX.modelranking}
\title{CausalKinetix.modelranking}
\usage{
CausalKinetiX.modelranking(D, times, env, target, models, pars)
}
\arguments{
\item{D}{data matrix. Should have dimension n x (L*d), where n is
the number of repetitions (over all experiments), L is the
number of time points and d is the number of predictor
variables.}

\item{times}{vector of length L specifying the time points at
which data was observed.}

\item{env}{integer vector of length n encoding to which experiment
each repetition belongs.}

\item{target}{integer specifing which variable is the target.}

\item{models}{list of models. Each model is specified by a list of
vectors specifiying the variables included in the interactions
of each term.}

\item{pars}{list of parameters.}
}
\value{
returns a vector with the same length as models containing
  the stability scores
}
\description{
Applies CausalKinetiX framework to rank a list models according to their stability
}
\details{
This function only scores specified models and does not include a variable ranking.
}
\examples{

x <- 4
}
\references{
Pfister, N., S. Bauer, J. Peters (2018).
Identifying Causal Structure in Large-Scale Kinetic Systems
ArXiv e-prints (arXiv:1810.11776).
}
\seealso{

}
\author{
Niklas Pfister, Stefan Bauer and Jonas Peters
}
